apps/tactics-finder/requirements.txt
tqdm
python-chess
stockfish
anytree
redis
requests
python-dotenv
watchdog
mypy